Renamed mqttService to mqttClient
